Proyecto primera version finalizada

+++ kart.js
@@ -14,11 +14,42 @@ function kart() {
         });
         div.innerHTML += `<p>Total: $${total.toFixed(2)}</p>`;
     }
+    div.innerHTML+=`<button onclick="deleatKart()" class="deleatButton">Deleat</button>`
     document.body.appendChild(div);
 }
 
 function deleatKart()
 {
-    let carrito=[];
-    localStorage.setItem("games", JSON.stringify(carrito));
+    let carrito = JSON.parse(localStorage.getItem("games"));
+    if (carrito && carrito.length > 0){
+    Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes!"
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your file has been deleted.",
+            icon: "success"
+          });
+          carrito=[];
+          localStorage.setItem("games", JSON.stringify(carrito));
+          setTimeout(() => {
+            location.reload();
+        }, 5000);
+        }
+      });
+     
+    }else
+    {
+        Swal.fire({
+            title: "!Kart already empty!",
+            icon: "warring"
+          });
+    }
 }
